Inquisitor Ordinator Fix
Allows him to actually use his armour, and provides him with the appropriate pre-blessed weapon. As was intended, but I missed it. Somehow noticed the dagger but not the blade.
If he was a Psydonite, he'd auto get the trait to wear his armour, but in this server's infinite wisdom, it was decided that they shouldn't be. For whatever reason.

Additionally, this provides them with the helmet and back selection of the Adjudicator, too.